Fixes issue #139
Making HttpServerResponse.close() idempotent and reducing two flushes to one for most RequestHandlers.
