+++ rx-netty/src/main/java/io/reactivex/netty/protocol/http/server/HttpServerResponse.java
@@ -70,7 +70,8 @@ public class HttpServerResponse<T> extends DefaultChannelWriter<T> {
         return nettyResponse.getStatus();
     }
 
-    public Observable<Void> close() {
+    @Override
+    public Observable<Void> _close() {
 
         writeHeadersIfNotWritten();
